Servlet and SQL: redefine max walking distance radius as an integer in meters

+++ sql_scripts/create_smartparking_users_table.sql
@@ -1,7 +1,7 @@
 CREATE table smartparking_users (
     id INTEGER NOT NULL AUTO_INCREMENT,         # egyedi azonosító
     password VARCHAR(50) NOT NULL,              # egylőre titkosítatlanul, ha belép, akkor bejelentkezésenként egyedi azonosítót generálok, ami majd a http kérésekben azonosítja (ezt is érdemes itt letárolni?)
-    search_range REAL DEFAULT 0.5,              # a maximum megengedett távolság, amin belül parkolóhely érdekli a felhasználót
+    search_range INTEGER DEFAULT 0.5,           # a maximum megengedett távolság méterben, amin belül parkolóhely érdekli a felhasználót
     lot_requests INTEGER DEFAULT 0,             # parkolóhely lekérdezések száma, statisztikához
     last_login BIGINT(20),                      # utolsó belépés, pl.: fontos lehet a munkamenet lejératának megállapításához, ha lejár, újra be kell jelentkeznie
     recommended_lots INTEGER DEFAULT 0,         # felhasználó által beküldött szabad helyek száma, statisztikához
